NODE-1978: Transactions v4 spec (#2755)

+++ lang/shared/src/main/scala/com/wavesplatform/lang/v1/traits/domain/TransferAttachment.scala
@@ -15,7 +15,7 @@ sealed trait TransferAttachment {
       case EmptyAttachment     => impl.unit
     }
 }
-case class IntValue(value: Int)         extends TransferAttachment
+case class IntValue(value: Long)         extends TransferAttachment
 case class BooleanValue(value: Boolean) extends TransferAttachment
 case class StringValue(value: String)   extends TransferAttachment
 case class ByteStrValue(value: ByteStr) extends TransferAttachment
